Use uint for ServiceStatus

+++ PIDataReaderSvc.cs
@@ -55,13 +55,13 @@ namespace PIDataReaderSvc {
 
 		[StructLayout(LayoutKind.Sequential)]
 		public struct ServiceStatus {
-			public long dwServiceType;
+			public uint dwServiceType;
 			public ServiceState dwCurrentState;
-			public long dwControlsAccepted;
-			public long dwWin32ExitCode;
-			public long dwServiceSpecificExitCode;
-			public long dwCheckPoint;
-			public long dwWaitHint;
+			public uint dwControlsAccepted;
+			public uint dwWin32ExitCode;
+			public uint dwServiceSpecificExitCode;
+			public uint dwCheckPoint;
+			public uint dwWaitHint;
 		};
 
 		[DllImport("advapi32.dll", SetLastError = true)]
